Make compiled Int 64-bit (i64) with 8-byte heap slots
The WASM backend's Int was i32, diverging from the interpreter's i64 and
capping integers at ~2.1B. This migrates the whole backend to a 64-bit Int:

- Scalars: Int is i64 (literals, arithmetic, comparisons, bitwise, unary), with
  three-way kind promotion so i64/i32/f64 operands mix safely.
- Heap: every list element, tuple/record field, and closure capture is now an
  8-byte slot holding a universal i64 representation (pointers zero-extended,
  floats bit-reinterpreted via to_slot/from_slot). Lists, mk{N}, push/concat/
  drop, pattern matching, for-loops, and the dict keys/values/pairs helpers all
  use the 8-byte layout. This also unblocks floats-in-collections (the old
  "Float elements/fields not compiled" guards are gone) and big ints inside
  concretely-typed lists/records.
- ABI: generic/type-variable values keep the i32 generic ABI; concrete i64 Int
  args/returns convert at call, Apply, Assign, and closure boundaries. Closures
  use the i32 ABI for params/result and narrow captures.
- Runtime: print_int and int_to_string are i64; index args (substring/at) and
  dict keys/values narrow to i32 at their helper boundaries; actor message
  fields read at the 8-byte stride.

The old i32-range literal guard and the float-in-collection rejections are
removed (those tests now assert the features work). New differential tests cover
big-int arithmetic past 2^31, big ints in lists, and floats in lists/tuples.

+++ src/runtime.rs
@@ -233,7 +233,7 @@ fn host_print(mut caller: Caller<'_, ActorState>, ptr: i32, len: i32) -> Result<
     Ok(())
 }
 
-fn host_print_int(caller: Caller<'_, ActorState>, n: i32) -> Result<()> {
+fn host_print_int(caller: Caller<'_, ActorState>, n: i64) -> Result<()> {
     println!("[actor {}] {n}", caller.data().id);
     caller.data().output.lock().unwrap().push(n.to_string());
     Ok(())
